LayoutTests/imported/w3c/web-platform-tests/navigation-timing/nav2_test_navigation_type_backforward-expected.txt

frame "<!--frame1-->" - has 1 onunload handler(s)
frame "<!--frame1-->" - has 1 onunload handler(s)
CONSOLE MESSAGE: line 32: TypeError: undefined is not an object (evaluating 'navigation_frame.performance.getEntriesByType("navigation")[0].type')
Description

This test validates the value of window.performance.getEntriesByType("navigation")[0].type with a forward and back navigation.

This page should be loaded with a yellow background frame below. It should turn green for a starting navigation, back to yellow for a back navigation and then back to green again for a forward navigation.

Along the navigation timeline the window.performance.getEntriesByType("navigation")[0].type is checked for back_forward.

This test passes if all of the checks to the window.performance.getEntriesByType("navigation")[0].type are correct throughout the navigation scenario and the frame below ends with a green background. Otherwise, this test fails.



Harness Error (FAIL), message = TypeError: undefined is not an object (evaluating 'navigation_frame.performance.getEntriesByType("navigation")[0].type')
